Pause Pairings and Resume Pairing functionality implementation (#4)
Resolving Req: 513531 - Allow users to pause and resume pairups - this merge enables for the user to pause and resume the matches. All of the wording from design spec has been implemented along with the necessary functionality. Users when added to the Team can now be able to opt out from the beginning if they do not wish to receive any matches